+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2021-09-08
+
+### Added
+- Add new resource `mx_rspamd_multimap_entry` to configure Rspamd multimap module.
+
+### Removed
+- `mx_rspamd` no longer supports `postmaster` and `postmaster_score` properties. Multimap rules are configured now using a separate resource.
+
 ## [0.3.5] - 2021-05-14
 ### Added
 - Rspamd configurable score for multimap rule (see `mx_rspamd` new resource property: `postmaster_score`).